deleted hash, moved test
