+++ src/passes/stringify-walker.h
@@ -13,7 +13,7 @@ struct StringifyWalker
   Module* wasm;
   std::queue<Expression**> queue;
 
-  void walkModule(Module* module);
+  static void walkModule(SubType* self, Module* module);
   static void scan(SubType* self, Expression** currp);
   static void addUniqueSymbol(SubType* self, Expression** currp);
   static void doVisitExpression(SubType* self, Expression** currp);
@@ -22,34 +22,6 @@ struct StringifyWalker
 private:
   static void handler(SubType* self, Expression**);
   static void deferredScan(SubType* self, Expression** currp);
-};
-
-struct HashStringifyWalker : public StringifyWalker<HashStringifyWalker> {
-
-  void walkModule(Module* module);
-  static void addUniqueSymbol(HashStringifyWalker* self, Expression** currp);
-  void visitExpression(Expression* curr);
-
-private:
-  // std::vector<uint64_t> string;
-  // uint64_t monotonic = 0;
-  //  Change key to Expression
-  //  std::unordered_map<Expression *, uint64_t> exprToCounter;
-  //  std::unordered_map<uint64_t, uint64_t> exprToCounter;
-
-  void addExpressionHash(Expression* curr, uint64_t hash);
-};
-
-struct TestStringifyWalker : public StringifyWalker<TestStringifyWalker> {
-
-  std::ostream& os;
-
-  TestStringifyWalker(std::ostream& os);
-
-  void walkModule(Module* module);
-  static void addUniqueSymbol(TestStringifyWalker* self, Expression** currp);
-  void visitExpression(Expression* curr);
-  void print(std::ostream& os);
 };
 
 } // namespace wasm
